Trait evm::evm::Finalize

pub trait Finalize {
    fn finalize<E: Ext>(self, ext: E) -> Result<FinalizationResult>;
}

Types that can be “finalized” using an EVM.

In practice, this is just used to define an inherent impl on Reult<GasLeft<'a>>.

Required methods

fn finalize<E: Ext>(self, ext: E) -> Result<FinalizationResult>

Consume the externalities, call return if necessary, and produce call result.