Module evm::executor::stack

A stack-based executor with customizable state. A memory-based state is provided, but can replaced by a custom implementation, for exemple one interacting with a database.

Structs

Accessed

Log

MemoryStackAccount

MemoryStackState

MemoryStackSubstate

PrecompileOutput

Data returned by a precompile on success.

StackExecutor

Stack-based executor.

StackSubstateMetadata

Enums

PrecompileFailure

Data returned by a precompile in case of failure.

StackExitKind

Traits

PrecompileSet

A set of precompiles. Checks of the provided address being in the precompile set should be as cheap as possible since it may be called often.

StackState

Type Definitions

PrecompileFn

Precompiles function signature. Expected input arguments are: