evm::executor::stack

Struct MemoryStackSubstate

pub struct MemoryStackSubstate<'config> { /* private fields */ }

Implementations

impl<'config> MemoryStackSubstate<'config>

pub fn new(metadata: StackSubstateMetadata<'config>) -> Self

pub fn logs(&self) -> &[Log]

pub fn logs_mut(&mut self) -> &mut Vec<Log>

pub fn metadata(&self) -> &StackSubstateMetadata<'config>

pub fn metadata_mut(&mut self) -> &mut StackSubstateMetadata<'config>

pub fn deconstruct<B: Backend>( self, backend: &B, ) -> (impl IntoIterator<Item = Apply<impl IntoIterator<Item = (H256, H256)>>>, impl IntoIterator<Item = Log>)

Deconstruct the memory stack substate, return state to be applied. Panic if the substate is not in the top-level substate.

pub fn enter(&mut self, gas_limit: u64, is_static: bool)

pub fn exit_commit(&mut self) -> Result<(), ExitError>

pub fn exit_revert(&mut self) -> Result<(), ExitError>

pub fn exit_discard(&mut self) -> Result<(), ExitError>

pub fn known_account(&self, address: H160) -> Option<&MemoryStackAccount>

pub fn known_basic(&self, address: H160) -> Option<Basic>

pub fn known_code(&self, address: H160) -> Option<Vec<u8>>

pub fn known_empty(&self, address: H160) -> Option<bool>

pub fn known_storage(&self, address: H160, key: H256) -> Option<H256>

pub fn known_original_storage(&self, address: H160) -> Option<H256>

pub fn is_cold(&self, address: H160) -> bool

pub fn is_storage_cold(&self, address: H160, key: H256) -> bool

pub fn deleted(&self, address: H160) -> bool

pub fn inc_nonce<B: Backend>( &mut self, address: H160, backend: &B, ) -> Result<(), ExitError>

pub fn set_storage(&mut self, address: H160, key: H256, value: H256)

pub fn reset_storage<B: Backend>(&mut self, address: H160, backend: &B)

pub fn log(&mut self, address: H160, topics: Vec<H256>, data: Vec<u8>)

pub fn set_deleted(&mut self, address: H160)

pub fn set_code<B: Backend>( &mut self, address: H160, code: Vec<u8>, backend: &B, )

pub fn transfer<B: Backend>( &mut self, transfer: Transfer, backend: &B, ) -> Result<(), ExitError>

pub fn withdraw<B: Backend>( &mut self, address: H160, value: U256, backend: &B, ) -> Result<(), ExitError>

pub fn deposit<B: Backend>(&mut self, address: H160, value: U256, backend: &B)

pub fn reset_balance<B: Backend>(&mut self, address: H160, backend: &B)

pub fn touch<B: Backend>(&mut self, address: H160, backend: &B)