Struct evm::executor::stack::StackExecutor

pub struct StackExecutor<'config, 'precompiles, S, P> { /* private fields */ }

Stack-based executor.

Implementations

impl<'config, 'precompiles, S: StackState<'config>, P: PrecompileSet> StackExecutor<'config, 'precompiles, S, P>

pub fn config(&self) -> &'config Config

Return a reference of the Config.

pub fn precompiles(&self) -> &'precompiles P

Return a reference to the precompile set.

pub fn new_with_precompiles(
state: S,
config: &'config Config,
precompile_set: &'precompiles P
) -> Self

Create a new stack-based executor with given precompiles.

pub fn state(&self) -> &S

pub fn state_mut(&mut self) -> &mut S

pub fn into_state(self) -> S

pub fn enter_substate(&mut self, gas_limit: u64, is_static: bool)

Create a substate executor from the current executor.

pub fn exit_substate(&mut self, kind: StackExitKind) -> Result<(), ExitError>

Exit a substate. Panic if it results an empty substate stack.

pub fn execute(&mut self, runtime: &mut Runtime<'_>) -> ExitReason

Execute the runtime until it returns.

pub fn gas(&self) -> u64

Get remaining gas.

pub fn transact_create(
&mut self,
caller: H160,
value: U256,
init_code: Vec<u8>,
gas_limit: u64,
access_list: Vec<(H160, Vec<H256>)>
) -> (ExitReason, Vec<u8>)

Execute a CREATE transaction.

pub fn transact_create2(
&mut self,
caller: H160,
value: U256,
init_code: Vec<u8>,
salt: H256,
gas_limit: u64,
access_list: Vec<(H160, Vec<H256>)>
) -> (ExitReason, Vec<u8>)

Execute a CREATE2 transaction.

pub fn transact_call(
&mut self,
caller: H160,
address: H160,
value: U256,
data: Vec<u8>,
gas_limit: u64,
access_list: Vec<(H160, Vec<H256>)>
) -> (ExitReason, Vec<u8>)

Execute a CALL transaction with a given caller, address, value and gas limit and data.

Takes in an additional access_list parameter for EIP-2930 which was introduced in the Ethereum Berlin hard fork. If you do not wish to use this functionality, just pass in an empty vector.

pub fn used_gas(&self) -> u64

Get used gas for the current executor, given the price.

pub fn fee(&self, price: U256) -> U256

Get fee needed for the current executor, given the price.

pub fn nonce(&self, address: H160) -> U256

Get account nonce.

pub fn create_address(&self, scheme: CreateScheme) -> H160

Get the create address from given scheme.

pub fn initialize_with_access_list(
&mut self,
access_list: Vec<(H160, Vec<H256>)>
)