Trait evm::executor::stack::PrecompileSet

pub trait PrecompileSet {
    fn execute(
        &self,
        address: H160,
        input: &[u8],
        gas_limit: Option<u64>,
        context: &Context,
        is_static: bool
    ) -> Option<Result<PrecompileOutput, PrecompileFailure>>;
    fn is_precompile(&self, address: H160) -> bool;
}

A set of precompiles. Checks of the provided address being in the precompile set should be as cheap as possible since it may be called often.

Required Methods

fn execute(
&self,
address: H160,
input: &[u8],
gas_limit: Option<u64>,
context: &Context,
is_static: bool
) -> Option<Result<PrecompileOutput, PrecompileFailure>>

Tries to execute a precompile in the precompile set. If the provided address is not a precompile, returns None.

fn is_precompile(&self, address: H160) -> bool

Check if the given address is a precompile. Should only be called to perform the check while not executing the precompile afterward, since execute already performs a check internally.