evm

Struct Machine

pub struct Machine { /* private fields */ }

Core execution layer for EVM.

Implementations

impl Machine

pub fn stack(&self) -> &Stack

Reference of machine stack.

pub fn stack_mut(&mut self) -> &mut Stack

Mutable reference of machine stack.

pub fn memory(&self) -> &Memory

Reference of machine memory.

pub fn memory_mut(&mut self) -> &mut Memory

Mutable reference of machine memory.

pub fn position(&self) -> &Result<usize, ExitReason>

Return a reference of the program counter.

pub fn new( code: Rc<Vec<u8>>, data: Rc<Vec<u8>>, stack_limit: usize, memory_limit: usize, ) -> Machine

Create a new machine with given code and data.

pub fn exit(&mut self, reason: ExitReason)

Explicit exit of the machine. Further step will return error.

pub fn inspect(&self) -> Option<(Opcode, &Stack)>

Inspect the machine’s next opcode and current stack.

pub fn return_value(&self) -> Vec<u8>

Copy and get the return value of the machine, if any.

pub fn run(&mut self) -> Capture<ExitReason, Opcode>

Loop stepping the machine, until it stops.

pub fn step(&mut self) -> Result<(), Capture<ExitReason, Opcode>>

Step the machine, executing one opcode. It then returns.