evm::executor

Module stack

A stack-based executor with customizable state. A memory-based state is provided, but can replaced by a custom implementation, for exemple one interacting with a database.

Structs

Accessed

Log

MemoryStackAccount

MemoryStackState

MemoryStackSubstate

PrecompileOutput

Data returned by a precompile on success.

StackExecutor

Stack-based executor.

StackSubstateMetadata

Enums

IsPrecompileResult

PrecompileFailure

Data returned by a precompile in case of failure.

StackExitKind

Traits

PrecompileHandle

Handle provided to a precompile to interact with the EVM.

PrecompileSet

A set of precompiles.

StackState

Type Aliases

PrecompileFn

Precompiles function signature. Expected input arguments are: